+++ core/app/[locale]/(default)/(auth)/register/page.tsx
@@ -8,7 +8,6 @@ import {
   formFieldTransformer,
   injectCountryCodeOptions,
 } from '~/data-transformers/form-field-transformer';
-import { getReCaptchaSettings } from '~/lib/recaptcha';
 import {
   CUSTOMER_FIELDS_TO_EXCLUDE,
   REGISTER_CUSTOMER_FORM_LAYOUT,
@@ -64,11 +63,6 @@ export default async function Register({ params }: Props) {
   const { addressFields, customerFields, countries, passwordComplexitySettings } =
     registerCustomerData;
 
-  const reCaptchaSettings = await getReCaptchaSettings();
-  const recaptchaEnabled =
-    (reCaptchaSettings?.isEnabledOnStorefront === true && Boolean(reCaptchaSettings?.siteKey)) ??
-    false;
-
   const fields = transformFieldsToLayout(
     [
       ...addressFields.map((field) => {
@@ -115,7 +109,6 @@ export default async function Register({ params }: Props) {
   return (
     <DynamicFormSection
       action={registerCustomer}
-      recaptchaEnabled={recaptchaEnabled}
       errorTranslations={{
         firstName: {
           invalid_type: t('FieldErrors.firstNameRequired'),
