+++ core/lib/recaptcha.ts
@@ -1,8 +1,14 @@
+import 'server-only';
+
 import { cache } from 'react';
 
 import { client } from '~/client';
 import { graphql } from '~/client/graphql';
 import { revalidate } from '~/client/revalidate-target';
+import type { ReCaptchaSettings } from './recaptcha/constants';
+
+export { RECAPTCHA_TOKEN_FORM_KEY } from './recaptcha/constants';
+export type { ReCaptchaSettings } from './recaptcha/constants';
 
 export const ReCaptchaSettingsQuery = graphql(`
   query ReCaptchaSettingsQuery {
@@ -18,13 +24,6 @@ export const ReCaptchaSettingsQuery = graphql(`
     }
   }
 `);
-
-export type ReCaptchaSettings = {
-  failedLoginLockoutDurationSeconds: number | null;
-  isEnabledOnCheckout: boolean;
-  isEnabledOnStorefront: boolean;
-  siteKey: string;
-};
 
 export const getReCaptchaSettings = cache(async (): Promise<ReCaptchaSettings | null> => {
   const { data } = await client.fetch({
@@ -45,6 +44,3 @@ export const getReCaptchaSettings = cache(async (): Promise<ReCaptchaSettings | 
     siteKey: reCaptcha.siteKey,
   };
 });
-
-/** FormData key used to pass the reCAPTCHA token from client to server actions */
-export const RECAPTCHA_TOKEN_FORM_KEY = 'recaptchaToken';
