+++ core/lib/recaptcha/constants.ts
@@ -1,0 +1,9 @@
+export type ReCaptchaSettings = {
+  failedLoginLockoutDurationSeconds: number | null;
+  isEnabledOnCheckout: boolean;
+  isEnabledOnStorefront: boolean;
+  siteKey: string;
+};
+
+/** FormData key used to pass the reCAPTCHA token from client to server actions */
+export const RECAPTCHA_TOKEN_FORM_KEY = 'recaptchaToken';
