+++ core/vibes/soul/sections/reviews/index.tsx
@@ -40,7 +40,6 @@ interface Props {
   }>;
   streamableProduct: Streamable<{ name: string }>;
   streamableUser: Streamable<{ email: string; name: string }>;
-  recaptchaEnabled?: boolean;
 }
 
 export function Reviews({
@@ -66,7 +65,6 @@ export function Reviews({
   streamableProduct,
   streamableImages,
   streamableUser,
-  recaptchaEnabled = false,
 }: Readonly<Props>) {
   return (
     <Stream fallback={<ReviewsSkeleton reviewsLabel={reviewsLabel} />} value={streamableReviews}>
@@ -86,7 +84,6 @@ export function Reviews({
               formTitleLabel={formTitleLabel}
               message={emptyStateMessage}
               productId={productId}
-              recaptchaEnabled={recaptchaEnabled}
               reviewsLabel={reviewsLabel}
               streamableImages={streamableImages}
               streamableProduct={streamableProduct}
@@ -140,7 +137,6 @@ export function Reviews({
                   formRatingLabel={formRatingLabel}
                   formReviewLabel={formReviewLabel}
                   formSubmitLabel={formSubmitLabel}
-                  recaptchaEnabled={recaptchaEnabled}
                   formTitleLabel={formTitleLabel}
                   productId={productId}
                   streamableImages={streamableImages}
@@ -205,7 +201,6 @@ export function ReviewsEmptyState({
   streamableProduct,
   streamableImages,
   streamableUser,
-  recaptchaEnabled = false,
 }: {
   message?: string;
   reviewsLabel?: string;
@@ -226,7 +221,6 @@ export function ReviewsEmptyState({
   }>;
   streamableProduct: Streamable<{ name: string }>;
   streamableUser: Streamable<{ email: string; name: string }>;
-  recaptchaEnabled?: boolean;
 }) {
   return (
     <StickySidebarLayout
@@ -256,7 +250,6 @@ export function ReviewsEmptyState({
           formSubmitLabel={formSubmitLabel}
           formTitleLabel={formTitleLabel}
           productId={productId}
-          recaptchaEnabled={recaptchaEnabled}
           streamableImages={streamableImages}
           streamableProduct={streamableProduct}
           streamableUser={streamableUser}
